Rewrite and rename base config
