+++ test/vanilla/.eslintrc.js
@@ -1,13 +1,14 @@
-const config = require('../../basic');
+/* eslint-disable */
+const config = require('../../vanilla');
 
 config.parserOptions = {
-  ecmaVersion: 9,
+  ecmaVersion: 2018,
   sourceType: 'module',
 };
+
 config.env = {
   'es2017': true,
   'browser': true,
-  'commonjs': true,
 };
 
 module.exports = config;
