+++ test/vanilla/comma-dangles.js
