+++ test/vanilla/no-var.js
