+++ test/vanilla/quotes.js
